ARTEMIS-2829 wrong return type for getDiskStoreUsage
Since getDiskStoreUsage on the ActiveMQServerControl is converting a
double to a long the value is always 0 in the management API. It should
return a double instead.

+++ artemis-core-client/src/main/java/org/apache/activemq/artemis/api/core/management/ActiveMQServerControl.java
@@ -445,10 +445,10 @@ public interface ActiveMQServerControl {
    long getAddressMemoryUsage();
 
    /**
-    * Returns the bytes used by the disk store
+    * Returns the percentage of total disk store use
     */
    @Attribute(desc = DISK_STORE_USAGE_DESCRIPTION)
-   long getDiskStoreUsage();
+   double getDiskStoreUsage();
 
    /**
     * Returns the memory used by all the addresses on broker as a percentage of the global-max-size
